Move orphaned scriptsstatusimport_test into install_third_party_libs_test
